# lib/api-zod/src/generated/types/extractedEnquiry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedEnquiryItem } from './extractedEnquiryItem';

/**
 * Structured enquiry produced by AI extraction. Missing information is null or empty; nothing is invented.
 */
export interface ExtractedEnquiry {
  companyName: string;
  contactName: string;
  phone: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  location: string | null;
  items: ExtractedEnquiryItem[];
  /** @nullable */
  notes: string | null;
  /** Questions the salesperson should ask because the enquiry did not state the information. */
  missingInformation: string[];
}